third_party/blink/web_tests/fast/dom/constructors-overriding-expected.txt

This page tests whether certain global constructors are read-only. Previous versions of WebKit incorrectly made them read-only.

If the test passes, you'll see a series of pass message below.

PASS: window.Image should be 1 and is.
PASS: window.Option should be 1 and is.
PASS: window.XMLHttpRequest should be 1 and is.
PASS: window.XMLSerializer should be 1 and is.
PASS: window.DOMParser should be 1 and is.
PASS: window.XSLTProcessor should be 1 and is.
